+++ AdColony/src/main/java/com/mopub/mobileads/AdColonyRewardedVideo.java
@@ -19,6 +19,7 @@ import com.mopub.common.DataKeys;
 import com.mopub.common.LifecycleListener;
 import com.mopub.common.MediationSettings;
 import com.mopub.common.MoPubReward;
+import com.mopub.common.Preconditions;
 import com.mopub.common.logging.MoPubLog;
 import com.mopub.common.util.Json;
 
@@ -36,6 +37,7 @@ import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOULD_REWARD;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_ATTEMPTED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_FAILED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_SUCCESS;
+import static com.mopub.mobileads.MoPubErrorCode.EXPIRED;
 
 public class AdColonyRewardedVideo extends CustomEventRewardedVideo {
 
@@ -48,8 +50,6 @@ public class AdColonyRewardedVideo extends CustomEventRewardedVideo {
     private AdColonyAdapterConfiguration mAdColonyAdapterConfiguration;
 
     private AdColonyInterstitial mAd;
-    @NonNull
-    private String mZoneId = AdColonyAdapterConfiguration.DEFAULT_ZONE_ID;
     private AdColonyListener mAdColonyListener;
     private AdColonyAdOptions mAdColonyAdOptions = new AdColonyAdOptions();
     private String mAdColonyClientOptions = "";
@@ -57,6 +57,15 @@ public class AdColonyRewardedVideo extends CustomEventRewardedVideo {
     @NonNull
     private String mAdUnitId = "";
     private boolean mIsLoading = false;
+
+    @NonNull
+    private static String mZoneId = AdColonyAdapterConfiguration.DEFAULT_ZONE_ID;
+
+    @NonNull
+    @Override
+    public String getAdNetworkId() {
+        return mZoneId;
+    }
 
     private AdColonyAppOptions mAdColonyAppOptions;
 
@@ -82,20 +91,22 @@ public class AdColonyRewardedVideo extends CustomEventRewardedVideo {
         return sLifecycleListener;
     }
 
-    @NonNull
-    @Override
-    public String getAdNetworkId() {
-        return mZoneId;
-    }
-
     @Override
     protected void onInvalidate() {
         mScheduledThreadPoolExecutor.shutdownNow();
         AdColonyInterstitial ad = sZoneIdToAdMap.get(mZoneId);
         if (ad != null) {
+            cleanUp(ad);
+        }
+    }
+
+    private static void cleanUp(AdColonyInterstitial ad) {
+        if (ad != null) {
             ad.destroy();
-            sZoneIdToAdMap.remove(mZoneId);
-            MoPubLog.log(CUSTOM, ADAPTER_NAME, "AdColony rewarded video destroyed");
+            if (sZoneIdToAdMap != null) {
+                sZoneIdToAdMap.remove(mZoneId);
+            }
+            MoPubLog.log(mZoneId, CUSTOM, ADAPTER_NAME, "AdColony rewarded video destroyed");
         }
     }
 
@@ -214,7 +225,7 @@ public class AdColonyRewardedVideo extends CustomEventRewardedVideo {
         AdColony.setRewardListener(mAdColonyListener);
         AdColony.requestInterstitial(mZoneId, mAdColonyListener, mAdColonyAdOptions);
         scheduleOnVideoReady();
-        MoPubLog.log(mZoneId, LOAD_ATTEMPTED, ADAPTER_NAME);
+        MoPubLog.log(getAdNetworkId(), LOAD_ATTEMPTED, ADAPTER_NAME);
     }
 
     private void abortRequestForIncorrectParameter(String parameterName) {
@@ -237,7 +248,7 @@ public class AdColonyRewardedVideo extends CustomEventRewardedVideo {
 
     @Override
     public void showVideo() {
-        MoPubLog.log(SHOW_ATTEMPTED, ADAPTER_NAME);
+        MoPubLog.log(getAdNetworkId(), SHOW_ATTEMPTED, ADAPTER_NAME);
         if (this.hasVideoAvailable()) {
             mAd.show();
         } else {
@@ -245,7 +256,7 @@ public class AdColonyRewardedVideo extends CustomEventRewardedVideo {
                     AdColonyRewardedVideo.class,
                     mZoneId,
                     MoPubErrorCode.NETWORK_NO_FILL);
-            MoPubLog.log(SHOW_FAILED, ADAPTER_NAME, MoPubErrorCode.NETWORK_NO_FILL.getIntCode(), MoPubErrorCode.NETWORK_NO_FILL);
+            MoPubLog.log(getAdNetworkId(), SHOW_FAILED, ADAPTER_NAME, MoPubErrorCode.NETWORK_NO_FILL.getIntCode(), MoPubErrorCode.NETWORK_NO_FILL);
         }
     }
 
@@ -288,12 +299,12 @@ public class AdColonyRewardedVideo extends CustomEventRewardedVideo {
                         @Override
                         public void run() {
                             if (hasVideoAvailable()) {
-                                MoPubLog.log(LOAD_SUCCESS, ADAPTER_NAME);
+                                MoPubLog.log(getAdNetworkId(), LOAD_SUCCESS, ADAPTER_NAME);
                                 MoPubRewardedVideoManager.onRewardedVideoLoadSuccess(
                                         AdColonyRewardedVideo.class,
                                         mZoneId);
                             } else {
-                                MoPubLog.log(LOAD_FAILED, ADAPTER_NAME, MoPubErrorCode.NETWORK_NO_FILL.getIntCode(), MoPubErrorCode.NETWORK_NO_FILL);
+                                MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME, MoPubErrorCode.NETWORK_NO_FILL.getIntCode(), MoPubErrorCode.NETWORK_NO_FILL);
                                 MoPubRewardedVideoManager.onRewardedVideoLoadFailure(
                                         AdColonyRewardedVideo.class,
                                         mZoneId,
@@ -318,29 +329,35 @@ public class AdColonyRewardedVideo extends CustomEventRewardedVideo {
     private static class AdColonyListener extends AdColonyInterstitialListener
             implements AdColonyRewardListener, CustomEventRewardedVideoListener {
         private AdColonyAdOptions mAdOptions;
+        private Boolean videoClosed = false;
 
         AdColonyListener(AdColonyAdOptions adOptions) {
             mAdOptions = adOptions;
+        }
+
+        @NonNull
+        private String getAdNetworkId() {
+            return mZoneId;
         }
 
         @Override
         public void onReward(@NonNull AdColonyReward a) {
             MoPubReward reward;
             if (a.success()) {
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "AdColonyReward name - " + a.getRewardName());
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "AdColonyReward amount - " + a.getRewardAmount());
+                MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "AdColonyReward name - " + a.getRewardName());
+                MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "AdColonyReward amount - " + a.getRewardAmount());
                 reward = MoPubReward.success(a.getRewardName(), a.getRewardAmount());
 
-                MoPubLog.log(SHOULD_REWARD, ADAPTER_NAME, a.getRewardAmount(), a.getRewardName());
-            } else {
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "AdColonyReward failed");
-                reward = MoPubReward.failure();
-            }
+                MoPubLog.log(getAdNetworkId(), SHOULD_REWARD, ADAPTER_NAME, a.getRewardAmount(), a.getRewardName());
 
-            MoPubRewardedVideoManager.onRewardedVideoCompleted(
-                    AdColonyRewardedVideo.class,
-                    a.getZoneID(),
-                    reward);
+                MoPubRewardedVideoManager.onRewardedVideoCompleted(
+                        AdColonyRewardedVideo.class,
+                        a.getZoneID(),
+                        reward);
+            } else {
+                MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "AdColonyReward failed");
+                closeAdColonyVideo(a.getZoneID());
+            }
         }
 
         @Override
@@ -350,20 +367,27 @@ public class AdColonyRewardedVideo extends CustomEventRewardedVideo {
 
         @Override
         public void onRequestNotFilled(@NonNull AdColonyZone zone) {
-            MoPubLog.log(CUSTOM, ADAPTER_NAME, "AdColony rewarded ad has no fill");
+            MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "AdColony rewarded ad has no fill");
             MoPubRewardedVideoManager.onRewardedVideoLoadFailure(
                     AdColonyRewardedVideo.class,
                     zone.getZoneID(),
                     MoPubErrorCode.NETWORK_NO_FILL);
-            MoPubLog.log(LOAD_FAILED, ADAPTER_NAME, MoPubErrorCode.NETWORK_NO_FILL.getIntCode(), MoPubErrorCode.NETWORK_NO_FILL);
+            MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME, MoPubErrorCode.NETWORK_NO_FILL.getIntCode(), MoPubErrorCode.NETWORK_NO_FILL);
         }
 
         @Override
         public void onClosed(@NonNull AdColonyInterstitial ad) {
-            MoPubLog.log(CUSTOM, ADAPTER_NAME, "Adcolony rewarded ad has been dismissed");
+            if (!videoClosed) {
+                closeAdColonyVideo(ad.getZoneID());
+            }
+        }
+
+        private void closeAdColonyVideo(String zoneId) {
+            videoClosed = true;
+            MoPubLog.log(zoneId, CUSTOM, ADAPTER_NAME, "AdColony rewarded ad has been dismissed");
             MoPubRewardedVideoManager.onRewardedVideoClosed(
                     AdColonyRewardedVideo.class,
-                    ad.getZoneID());
+                    zoneId);
         }
 
         @Override
@@ -371,12 +395,22 @@ public class AdColonyRewardedVideo extends CustomEventRewardedVideo {
             MoPubRewardedVideoManager.onRewardedVideoStarted(
                     AdColonyRewardedVideo.class,
                     ad.getZoneID());
-            MoPubLog.log(SHOW_SUCCESS, ADAPTER_NAME);
+            MoPubLog.log(getAdNetworkId(), SHOW_SUCCESS, ADAPTER_NAME);
         }
 
         @Override
         public void onExpiring(@NonNull AdColonyInterstitial ad) {
-            AdColony.requestInterstitial(ad.getZoneID(), ad.getListener(), mAdOptions);
+            Preconditions.checkNotNull(ad);
+
+            final String expiredZoneId = ad.getZoneID();
+            MoPubLog.log(expiredZoneId, CUSTOM, ADAPTER_NAME, "Expiring unused " +
+                    "AdColony Rewarded Video ad due to AdColony's 30-minute expiration policy.");
+            MoPubRewardedVideoManager.onRewardedVideoPlaybackError(AdColonyRewardedVideo.class,
+                    expiredZoneId, EXPIRED);
+            MoPubLog.log(expiredZoneId, LOAD_FAILED, ADAPTER_NAME,
+                    MoPubErrorCode.EXPIRED.getIntCode(), MoPubErrorCode.EXPIRED);
+
+            cleanUp(ad);
         }
 
         @Override
@@ -384,7 +418,7 @@ public class AdColonyRewardedVideo extends CustomEventRewardedVideo {
             MoPubRewardedVideoManager.onRewardedVideoClicked(
                     AdColonyRewardedVideo.class,
                     ad.getZoneID());
-            MoPubLog.log(CLICKED, ADAPTER_NAME);
+            MoPubLog.log(getAdNetworkId(), CLICKED, ADAPTER_NAME);
         }
     }