+++ AdMob/src/main/java/com/mopub/mobileads/GooglePlayServicesInterstitial.java
@@ -1,12 +1,10 @@
 package com.mopub.mobileads;
 
 import android.content.Context;
-import android.os.Bundle;
 import android.text.TextUtils;
 
 import androidx.annotation.NonNull;
 
-import com.google.ads.mediation.admob.AdMobAdapter;
 import com.google.android.gms.ads.AdListener;
 import com.google.android.gms.ads.AdRequest;
 import com.google.android.gms.ads.InterstitialAd;
@@ -14,6 +12,7 @@ import com.google.android.gms.ads.MobileAds;
 import com.google.android.gms.ads.RequestConfiguration;
 import com.mopub.common.logging.MoPubLog;
 
+import java.util.Collections;
 import java.util.Map;
 
 import static com.google.android.gms.ads.RequestConfiguration.TAG_FOR_CHILD_DIRECTED_TREATMENT_FALSE;
@@ -28,6 +27,7 @@ import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_SUCCESS;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_ATTEMPTED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_FAILED;
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_SUCCESS;
+import static com.mopub.mobileads.GooglePlayServicesAdapterConfiguration.forwardNpaIfSet;
 
 public class GooglePlayServicesInterstitial extends CustomEventInterstitial {
     /*
@@ -57,6 +57,7 @@ public class GooglePlayServicesInterstitial extends CustomEventInterstitial {
             final Map<String, Object> localExtras,
             final Map<String, String> serverExtras) {
 
+        MobileAds.initialize(context);
         setAutomaticImpressionAndClickTracking(false);
 
         mInterstitialListener = customEventInterstitialListener;
@@ -81,7 +82,7 @@ public class GooglePlayServicesInterstitial extends CustomEventInterstitial {
         mGoogleInterstitialAd.setAdListener(new InterstitialAdListener());
         mGoogleInterstitialAd.setAdUnitId(mAdUnitId);
 
-        final AdRequest.Builder builder = new AdRequest.Builder();
+        AdRequest.Builder builder = new AdRequest.Builder();
         builder.setRequestAgent("MoPub");
 
         // Publishers may append a content URL by passing it to the MoPubInterstitial.setLocalExtras() call.
@@ -91,18 +92,16 @@ public class GooglePlayServicesInterstitial extends CustomEventInterstitial {
             builder.setContentUrl(contentUrl);
         }
 
-        // Publishers may request for test ads by passing test device IDs to the MoPubInterstitial.setLocalExtras() call.
-        final String testDeviceId = (String) localExtras.get(TEST_DEVICES_KEY);
-
-        if (!TextUtils.isEmpty(testDeviceId)) {
-            builder.addTestDevice(testDeviceId);
-        }
-
-        // Consent collected from the MoPub’s consent dialogue should not be used to set up
-        // Google's personalization preference. Publishers should work with Google to be GDPR-compliant.
         forwardNpaIfSet(builder);
 
         final RequestConfiguration.Builder requestConfigurationBuilder = new RequestConfiguration.Builder();
+
+        // Publishers may request for test ads by passing test device IDs to the MoPubView.setLocalExtras() call.
+        final String testDeviceId = (String) localExtras.get(TEST_DEVICES_KEY);
+
+        if (!TextUtils.isEmpty(testDeviceId)) {
+            requestConfigurationBuilder.setTestDeviceIds(Collections.singletonList(testDeviceId));
+        }
 
         // Publishers may want to indicate that their content is child-directed and forward this
         // information to Google.
@@ -150,16 +149,6 @@ public class GooglePlayServicesInterstitial extends CustomEventInterstitial {
             if (mInterstitialListener != null) {
                 mInterstitialListener.onInterstitialFailed(MoPubErrorCode.NETWORK_NO_FILL);
             }
-        }
-    }
-
-    private void forwardNpaIfSet(AdRequest.Builder builder) {
-
-        // Only forward the "npa" bundle if it is explicitly set. Otherwise, don't attach it with the ad request.
-        final Bundle npaBundle = GooglePlayServicesAdapterConfiguration.getNpaBundle();
-
-        if (npaBundle != null && !npaBundle.isEmpty()) {
-            builder.addNetworkExtrasBundle(AdMobAdapter.class, npaBundle);
         }
     }