+++ AdMob/CHANGELOG.md
@@ -1,4 +1,18 @@
 ## Changelog
+  * 19.1.0.0
+    * This version of the adapters has been certified with AdMob 19.1.0 and MoPub 5.12.0.
+    * Remove calls to `MobileAds.initialize(Context context, String applicationCode)` (deprecated by Google) in favor of `MobileAds.initialize(Context context)`.
+
+  * 19.0.1.0
+    * This version of the adapters has been certified with AdMob 19.0.1 and MoPub 5.11.1.
+
+  * 19.0.0.0
+    * This version of the adapters has been certified with AdMob 19.0.0.
+    * Deprecated `AdRequest.Builder.addTestDevice()` in favor of `RequestConfiguration.Builder.setTestDeviceIds()`.
+
+  * 18.3.0.3
+    * MoPub now collects GDPR consent on behalf of Google.
+
   * 18.3.0.2
     * Log the AdMob ad unit ID in ad lifecycle events.