+++ FacebookAudienceNetwork/CHANGELOG.md
@@ -1,4 +1,16 @@
 ## Changelog
+ * 5.8.0.0
+   * This version of the adapters has been certified with Facebook Audience Network 5.8.0 and MoPub 5.11.1.
+
+ * 5.7.1.1
+   * Fix duplicate firing of `onRewardedVideoClosed()`. 
+
+ * 5.7.1.0
+   * This version of the adapters has been certified with Facebook Audience Network 5.7.1.
+
+ * 5.7.0.0
+   * This version of the adapters has been certified with Facebook Audience Network 5.7.0.
+
  * 5.6.1.0
    * This version of the adapters has been certified with Facebook Audience Network 5.6.1.