+++ IronSource/CHANGELOG.md
@@ -1,4 +1,21 @@
 ## Changelog
+ * 6.16.0.0
+    * This version of the adapters has been certified with ironSource 6.16.0 and MoPub SDK 5.12.0.
+    * Support for Android Archive (AAR) version of the ironSource SDK.
+
+ * 6.15.0.1.0
+    * This version of the adapters has been certified with ironSource 6.15.0.1 and MoPub SDK 5.11.1.
+
+ * 6.14.0.1.0
+    * This version of the adapters has been certified with ironSource 6.14.0.1.
+
+ * 6.14.0.0
+    * This version of the adapters has been certified with ironSource 6.14.0.
+
+ * 6.13.0.1.0
+    * This version of the adapters has been certified with ironSource 6.13.0.1.
+    * Log the ironSource instance id in ad lifecycle events, and improve error console logs.
+
  * 6.13.0.0
     * This version of the adapters has been certified with ironSource 6.13.0.