+++ Verizon/CHANGELOG.md
@@ -1,4 +1,15 @@
 ## Changelog
+  * 1.5.0.0
+    * Compress the Advanced Bidding token to adhere to MoPub's spec.
+    * This version of the adapters has been certified with Verizon 1.5.0 and MoPub 5.11.1.
+
+  * 1.4.0.0
+    * Update Advanced Bidding API
+    * This version of the adapters has been certified with Verizon 1.4.0.
+
+  * 1.3.1.0
+    * This version of the adapters has been certified with Verizon 1.3.1.
+  
   * 1.3.0.0
     * This version of the adapters has been certified with Verizon 1.3.0.
     * Remove Advanced Bidding token generation logic from the adapters. The equivalent logic will be added to the Verizon SDK.